# API-SPECIFICATION.yml
openapi: 3.0.3
info:
  title: ArtiCurated Order Management API
  description: |
    RESTful API for managing orders and returns in the ArtiCurated boutique marketplace.

    ## Key Features
    - Deterministic state machine for order lifecycle management
    - Multi-step returns workflow with manager approval
    - Comprehensive audit trail for all state transitions
    - Asynchronous background processing for invoices and refunds

    ## State Machines

    ### Order States
    - `PENDING_PAYMENT` → `PAID` → `PROCESSING_IN_WAREHOUSE` → `SHIPPED` → `DELIVERED`
    - `PENDING_PAYMENT` or `PAID` can transition to `CANCELLED`

    ### Return States
    - `REQUESTED` → `APPROVED` → `IN_TRANSIT` → `RECEIVED` → `COMPLETED`
    - `REQUESTED` can transition to `REJECTED` (terminal)

  version: 1.0.0
  contact:
    name: ArtiCurated Engineering
    email: engineering@articurated.com
servers:
  - url: http://localhost:8000/api/v1
    description: Development server
  - url: https://api.articurated.com/v1
    description: Production server
...
paths:
...
  /returns:
    post:
      tags: [Returns]
      summary: Initiate return request
      description: |
        Create a return request for a delivered order.

        **Requirements:**
        - Order must be in DELIVERED state
        - Within 30 days of delivery
        - Return reason required
      operationId: createReturn
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReturnCreate"
            example:
              order_id: "550e8400-e29b-41d4-a716-446655440000"
              reason: "Item damaged during shipping"
              customer_notes: "Box was crushed, vase has visible cracks"
      responses:
        "201":
          description: Return request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReturnResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Order not eligible for return
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: "RETURN_NOT_ALLOWED"
                  message: "Returns only allowed for DELIVERED orders"
                  details:
                    order_status: "SHIPPED"

    get:
      tags: [Returns]
      summary: List returns
      description: Retrieve paginated list of return requests
      operationId: listReturns
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/ReturnStatus"
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: List of returns
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/ReturnResponse"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
  /returns/{return_id}:
    get:
      tags: [Returns]
      summary: Get return details
      description: Retrieve complete return information
      operationId: getReturn
      parameters:
        - $ref: "#/components/parameters/ReturnId"
      responses:
        "200":
          description: Return details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReturnResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /returns/{return_id}/approve:
    patch:
      tags: [Returns]
      summary: Approve return request
      description: |
        Manager approves return request.
        Transitions state from REQUESTED to APPROVED.

        **Requirements:**
        - Manager notes required
      operationId: approveReturn
      parameters:
        - $ref: "#/components/parameters/ReturnId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                manager_notes:
                  type: string
                  maxLength: 1000
              required: [manager_notes]
            example:
              manager_notes: "Approved - valid damage claim with photo evidence"
      responses:
        "200":
          description: Return approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReturnResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/InvalidStateTransition"
  /returns/{return_id}/reject:
    patch:
      tags: [Returns]
      summary: Reject return request
      description: |
        Manager rejects return request.
        Transitions state from REQUESTED to REJECTED (terminal).

        **Requirements:**
        - Manager notes required
        - Rejection reason required
      operationId: rejectReturn
      parameters:
        - $ref: "#/components/parameters/ReturnId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                manager_notes:
                  type: string
                  maxLength: 1000
                rejection_reason:
                  type: string
                  enum:
                    [
                      damage_not_covered,
                      policy_violation,
                      outside_window,
                      fraudulent,
                    ]
              required: [manager_notes, rejection_reason]
            example:
              manager_notes: "Return window expired, order delivered 45 days ago"
              rejection_reason: "outside_window"
      responses:
        "200":
          description: Return rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReturnResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/InvalidStateTransition"
  /returns/{return_id}/state:
    patch:
      tags: [Returns]
      summary: Update return state
      description: |
        Transition return to new state (IN_TRANSIT, RECEIVED, COMPLETED).

        **Triggers:**
        - COMPLETED → Triggers refund processing background job
      operationId: updateReturnState
      parameters:
        - $ref: "#/components/parameters/ReturnId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StateTransition"
      responses:
        "200":
          description: State transition successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReturnResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/InvalidStateTransition"
  /returns/{return_id}/history:
    get:
      tags: [Returns]
      summary: Get return state history
      description: Retrieve complete audit trail of return state transitions
      operationId: getReturnHistory
      parameters:
        - $ref: "#/components/parameters/ReturnId"
      responses:
        "200":
          description: Return state history
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/StateHistoryRecord"
        "404":
          $ref: "#/components/responses/NotFound"

  # ==================== HEALTH & MONITORING ====================
...
components:
  parameters:
...
    ReturnId:
      name: return_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Unique return identifier
  schemas:
...
    ReturnStatus:
      type: string
      enum:
        - REQUESTED
        - APPROVED
        - REJECTED
        - IN_TRANSIT
        - RECEIVED
        - COMPLETED
      description: Return workflow states
    ActorType:
      type: string
      enum:
        - USER
        - SYSTEM
      description: Type of actor performing state transition

    # ==================== ORDER SCHEMAS ====================
...
    ReturnCreate:
      type: object
      required:
        - order_id
        - reason
      properties:
        order_id:
          type: string
          format: uuid
        reason:
          type: string
          maxLength: 1000
        customer_notes:
          type: string
          maxLength: 2000
    ReturnResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        order_id:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/ReturnStatus"
        reason:
          type: string
        customer_notes:
          type: string
          nullable: true
        manager_notes:
          type: string
          nullable: true
        rejection_reason:
          type: string
          nullable: true
        refund_amount:
          type: number
          format: decimal
        refund_transaction_id:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        approved_at:
          type: string
          format: date-time
          nullable: true
        rejected_at:
          type: string
          format: date-time
          nullable: true
        completed_at:
          type: string
          format: date-time
          nullable: true

    # ==================== STATE TRANSITION ====================
    StateTransition:
      type: object
      required:
        - new_state
      properties:
        new_state:
          type: string
        reason:
          type: string
          maxLength: 500
          nullable: true
    StateHistoryRecord:
      type: object
      properties:
        id:
          type: string
          format: uuid
        entity_type:
          type: string
          enum: [ORDER, RETURN]
        entity_id:
          type: string
          format: uuid
        previous_state:
          type: string
        new_state:
          type: string
        actor_id:
          type: string
          format: uuid
          nullable: true
        actor_type:
          $ref: "#/components/schemas/ActorType"
        trigger:
          type: string
          example: "API_CALL"
        metadata:
          type: object
          additionalProperties: true
        ip_address:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time

    # ==================== COMMON SCHEMAS ====================
    Pagination:
      type: object
      properties:
        limit:
          type: integer
        offset:
          type: integer
        total:
          type: integer
        has_more:
          type: boolean
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: "INVALID_STATE_TRANSITION"
            message:
              type: string
              example: "Cannot transition from SHIPPED to PAID"
            details:
              type: object
              additionalProperties: true
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: "BAD_REQUEST"
              message: "Invalid request parameters"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: "NOT_FOUND"
              message: "Order not found"
...
    InvalidStateTransition:
      description: Invalid state transition attempted
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: "INVALID_STATE_TRANSITION"
              message: "Cannot transition from REQUESTED to COMPLETED"
              details:
                current_state: "REQUESTED"
                requested_state: "COMPLETED"
                allowed_transitions: ["APPROVED", "REJECTED"]
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication
security:
  - ApiKeyAuth: []